[JNI] Use a read/write lock pattern in Rmm.class (#20521)
Closes https://github.com/rapidsai/cudf/issues/20520

This PR is adding a read/writer lock pattern to Rmm java, instead of relying on the whole-class static `synchronized` statement. This allows us to have multiple readers that only use the reader lock. Much of the improvement is actually external to this file, so I'll reference a spark-rapids-jni PR soon that uses the locks, showing why we need it.

Authors:
  - Alessandro Bellina (https://github.com/abellina)

Approvers:
  - Robert (Bobby) Evans (https://github.com/revans2)

URL: https://github.com/rapidsai/cudf/pull/20521