Capture query plan in pds-h output
Currently, pds-h benchmarks will optionally print out the query plan if
`--explain` is passed. With this PR, we'll also persist the query plan
in the `pdsh_results.jsonl` file. This will facilitate downstream
analysis.

To make it easier for downstream tools to work with, we persist a
structured version of the query plan, rather than the tree-like textual
representation.