Skip flaky upstream polars rolling test (#20552)
Admin merging based on discussion with the `build-eng` team. There was one unexpected failure but that looked to be related to transitory network issues vs code related to this PR so moving forward with the admin merge. 

## Description
<!-- Provide a standalone description of changes in this PR. -->
<!-- Reference any issues closed by this PR with "closes #1234". -->
<!-- Note: The pull request title will be included in the CHANGELOG. -->
Example of the test failing:
https://github.com/rapidsai/cudf/actions/runs/19144899197/job/54719782473

xref #20551 
## Checklist
- [ ] I am familiar with the [Contributing
Guidelines](https://github.com/rapidsai/cudf/blob/HEAD/CONTRIBUTING.md).
- [ ] New or existing tests cover these changes.
- [ ] The documentation is up to date with these changes.