Add pytest stubs and remove ujson usage (#20560)
pytest stubs are just added for completeness and didn't require any fixes. I removed ujson usage because we never test it to validate that it actually works for us. If we find that this meaningfully slows down particular use cases we can easily add it back (with proper testing), although at this point the better option would be to migrate to orjson, which is both faster and more correct.

Contributes to #11661 
Split out from #20547

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)
  - Matthew Roeschke (https://github.com/mroeschke)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/20560