[FEA] Optimize JIT Filter for Low-Selectivity (#20222)
Filters are typically performed by generating a Boolean mask from a set of columns.
The boolean mask is then converted into a gather mask, which is then used to copy from the source tables.
What we tried to do differently in JIT filters was to prevent the materialisation of a gather mask and get a directly usable filter representation from the predicate evaluation stage.
We used a selection representation where the value is either the current *selected* index or -1 (not selected). We then performed a transformed *copy_if* on the result. This was significantly faster than `apply_boolean_mask` (2-4x) as we didn't need to materialize the gather mask.
*copy_if* has the pathological case that happens when it selects a few values (<0.1%) on a large table because it needs to read the entire length of the selection mask, thus consuming more bandwidth. The effect also compounds when performing multiple copies. This ends up making it slower for low selectivity compared to `apply_boolean_mask`, which uses a compacted index (gather map) and a gather.
Whilst compacting the selection indices in the filter kernel would be possible, it would likely be inefficient due to the synchronization complexity and intermediates that would need to be materialized (prefix-scan).
I have made the filter kernel to use `apply_boolean_mask` instead.


Closes https://github.com/rapidsai/cudf/issues/20191

Authors:
  - Basit Ayantunde (https://github.com/lamarrr)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Karthikeyan (https://github.com/karthikeyann)

URL: https://github.com/rapidsai/cudf/pull/20222