Preprocessing offsets for Parquet non-dictionary string columns (#20430)
Contributes to #18350

For non-dictionary string columns, the strings are laid out in memory as: 4-byte length, string data ..., 4-byte length, string data, ... With this layout we cannot read the lengths in parallel, as we don't know where in global memory the next length value is without knowing the previous value. This causes the dominant bottleneck for non-dictionary, non-Fixed-Length-Byte-Array (all string lengths are the same) string columns, taking six times as long as the rest of the decode. 

This PR moves the reading of the string lengths out of the main decode kernel and into a preprocessing kernel.  In this kernel we allocate a small buffer of shared memory that all threads in the warp can cooperatively prefetch into from global memory.  Then when thread 0 extracts the string lengths, it is doing so from shared memory instead of global memory so we spend much less time bottlenecked on a single thread.  

However if the average string length is large, prefetching a buffer into shared memory is slower.  If only a few strings fit into the buffer, then we spend too much time prefetching data that we mostly don't need. Benchmarking showed that this performance dropoff was around fitting 16 length-64 strings into the 1KB shared memory buffer.  Thus above this cutoff we revert to a fully single-threaded, non-buffering algorithm. For strings lengths 128+ there was no performance change. 

There are several advantages to preprocessing instead of doing this in the main decode kernel.  First of all, by having only one warp per page instead of 4 warps per page in the decode kernel, we can go much wider, and not so many threads are bottlenecked.  But primarily we have significantly less shared memory and register pressure, and can safely throw a lot more memory at the prefetch buffer without reducing occupancy. 

The performance gains:
- Flat non-chunked: 20% faster
- Struct non-chunked: 25% faster
- List<string> non-chunked: 10% faster
- Chunked: 15% faster for all types. 

I attempted double-buffering the data, by having one warps filling half of the prefetch buffer while thread 0 on the other warp read from the other half, but this was slower than single-buffering. Comparing, with two warps (64 threads) and 2kb of shared memory you can do either:
- **Single buffering:** 2 pages at once, all 64 threads prefetch, but when buffers full 62 threads wait while the thread 0's extract lengths from shared memory
- **Double buffering:** 1 page at once, 32 threads continuously prefetch, 1 thread extracts lengths from other buffer, 31 threads do absolutely nothing

The increased speed from single buffering is because the 62 threads are probably not waiting very long for the thread 0's to extract the lengths, so mostly all 64 threads prefetch, instead of just 32 in the double-buffering scheme. 

Instead of saving both string lengths and offsets as before, only offsets are saved now, as the string lengths can be extracted from them. Also, the string lengths of for fixed-length-byte-array columns are known from the column data header, so they aren't saved at all. Finally, by doing this preprocessing, I was able to free some shared memory from the main strings decode kernel due to buffers no longer being needed.

Authors:
  - Paul Mattione (https://github.com/pmattione-nvidia)
  - Muhammad Haseeb (https://github.com/mhaseeb123)

Approvers:
  - https://github.com/nvdbaranec
  - Muhammad Haseeb (https://github.com/mhaseeb123)

URL: https://github.com/rapidsai/cudf/pull/20430