Optimize row mask computation for single filter column (#20335)
This PR optimizes the computation of row mask in case of single filter column in the experimental parquet reader. This is done by first computing the `cudf::compute_column` on a column containing page-level stats (one row per page) and then expanding boolean results to row level instead of the other way around. 

The advantage is two-fold:
1. `cudf::compute_column` operates on a smaller column
2. It's cheaper to expand a BOOL8 column from page level to row level as compared to expanding a numeric or especially string column to row level first before even applying a `cudf::compute_column`. We can't unfortunately avoid this if there are multiple filter columns as page boundaries don't align across columns 

Performance Improvement: While we don't yet have specific benchmarks for this, we did see a positive performance benefit in `hybrid_scan_io` example with this PR (still usually not enough to offset the cost of this filter but not too punishing anymore either).

Authors:
  - Muhammad Haseeb (https://github.com/mhaseeb123)

Approvers:
  - Vukasin Milovanovic (https://github.com/vuule)
  - https://github.com/nvdbaranec

URL: https://github.com/rapidsai/cudf/pull/20335