DatetimeIndex.serialize() is msgpack serializable
https://github.com/rapidsai/cudf/pull/20709 and
https://github.com/rapidsai/cudf/pull/18778 recently changed
DatetimeIndex to preserve `freq` in more places.

This exposed an issue in how we serialize the properties of a DatetimeIndex:
a `DateOffset` instance isn't serializable by msgpack. Distributed
msgpack serialize the headers of messages, and so putting the
`DateOffset` freq there caused issues.

Instead of inserting the `DateOffset` itself, we insert the `kwds` that
can be used to reconstruct it when deserializing.

Closes https://github.com/rapidsai/dask-upstream-testing/issues/88