Always delay CUDA Array Interface pointer access (#20719)
The current approach for accessing the pointer underlying an object is only valid for the original cuDF buffer model where data is never migrated. Both of our newer buffer models (copy-on-write and spilling) violate this assumption. That means that statically capturing the data pointer (specifically via the CAI since that is what our model is currently built around, but also via any other means) is invalid. In the longer term we will likely need to restructure our interaction across the cudf/pylibcudf boundary to handle this lazy access in a cleaner and more centralized manner, but for now this PR patches the outstanding issues using a few minimal CAI wrapper classes. In general we should assume that all uses of `types.SimpleNamespace` for capturing the CAI are suspect. The only one that I've left intact is the one that is directly taking input from external CAI-supporting objects since in those cases we know that external objects don't move the pointer out from underneath the way that we do.

Contributes to #19946.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Matthew Roeschke (https://github.com/mroeschke)

URL: https://github.com/rapidsai/cudf/pull/20719