The rapidsmpf runtime must use the rapidsmpf Pool
This updates our config to require using the rapidsmpf CUDA Stream Pool
with the rapidsmpf runtime. As we move pieces of the IR execution into
native rapidsmpf IR nodes (e.g. rapidsmpf's read_parquet), which
necessarily have to occur on a stream from the rapidsmpf stream pool,
it seems like the complexity of mixing multiple stream sources is
not worth any potential benefits.