Test copy-on-write in CI (#20745)
This PR selects a subset of the conda Python jobs to test with copy-on-write on. This bifurcation will hopefully not be too long-lived, so I have simply selected our current latest Python version under the hope that by the time we drop Python 3.13 support copy-on-write will be something we can largely assume to be adopted everywhere since it will have been the pandas default for a long time by then.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Matthew Roeschke (https://github.com/mroeschke)
  - James Lamb (https://github.com/jameslamb)

URL: https://github.com/rapidsai/cudf/pull/20745