Avoid generating empty `TableChunks` in streaming scan nodes (#20815)
We are currently pushing an empty TableChunk into the `scan_node` output channel when ranks are not assigned to any data. This should not be necessary and is technically a "bug" in the current implementation. Nodes should be able to drain immediately when they have no data to process.

This PR makes changes to several node implementations to handle the "empty-channel" case.

Authors:
  - Richard (Rick) Zamora (https://github.com/rjzamora)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Tom Augspurger (https://github.com/TomAugspurger)

URL: https://github.com/rapidsai/cudf/pull/20815