Fix some gtests to not assume dictionary keys order (#20827)
Fixes some gtests that are assuming keys are sorted (or predictably ordered) by changing the logic to check a decoded version against an expected result instead. The fixed tests now pass regardless of the order of the keys.

This is part of a larger effort to support unsorted keys in dictionary type columns.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Tianyu Liu (https://github.com/kingcrimsontianyu)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/20827