Parquet decode: Skip up to first_row for non-lists (#20835)
This changes the parquet decode to skip decoding work up to first_row for non-lists.  We were already skipping ahead for lists, which was easy to do because there is a lot of list preprocessing that tells us exactly how far to skip ahead.  

We don't have that for non-lists with nulls, so we have to do some work.  Specifically, if we skip N rows, we don't know how many valids we are skipping so we have to loop through and figure that out.  This loop (with the call to skip_validity_and_row_indices_nonlist()) is basically the same as the normal decode processing with all of the extra superfluous code/work stripped out.  We then can skip ahead with the dictionary and bool stream decode as before. 

This does not immediately improve performance, as the bottlenecks are elsewhere.  However, skipping ahead simplifies the rest of the code, removing weird corner cases (see removed if-checks) in the code that were necessary without the skipping.  This will make further simplifications and optimizations to the code easier later on.

Authors:
  - Paul Mattione (https://github.com/pmattione-nvidia)

Approvers:
  - Muhammad Haseeb (https://github.com/mhaseeb123)
  - https://github.com/nvdbaranec

URL: https://github.com/rapidsai/cudf/pull/20835