Reader and writer for a simple CudfTable format (#20811)
Closes https://github.com/rapidsai/cudf/issues/20302

Adds reader and writer for a format based on `cudf::pack` and `cudf::unpack`. The format does not encode or compress data.
The reader is zero-copy when the packed data is already on device, so the returned type is not `cudf::table`. Instead, a `packed_table` object that owns the device memory and exposes a `cudf::table_view` is returned.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Yunsong Wang (https://github.com/PointKernel)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/20811