Simplify categorical column (#20942)
This PR makes CategoricalColumn.codes a trivial access to `self.children[0]`, so there is no longer a separate `self._codes` attribute storing a masked version of the child column. With that change, we can also remove the need for `to_pylibcudf` to support the `use_base` parameter, which was only false to support setting the mask on the categorical codes due to the separation between the codes' mask and the categorical's mask. `to_pylibcudf` also no longer needs specialized handling for categorical columns since a simple override in CategoricalColumn is sufficient. Additionally, we can remove multiple overrides of parent methods from CategoricalColumn that only served to maintain the codes.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Matthew Roeschke (https://github.com/mroeschke)

URL: https://github.com/rapidsai/cudf/pull/20942