Add context manager to control access mode (#20952)
This PR introduces a new context manager to control the access mode of a buffer's pointer. This approach replaces the `get_ptr(mode=*)` approach, which only works as long as all pointer access occur within cudf itself. Since all computations ultimately happen within pylibcudf and we may need to support other libraries as well, a context manager like this provides a more flexible approach with much stronger correctness guarantees, which is particularly important now that copy-on-write is becoming the default and we need spilling to interoperate safely with CoW. In this PR the context manager is optional and ptr access default to read, but in a follow-up PR the context manager will be made mandatory (possibly with an escape hatch). That will be a much more wide-ranging but fairly rote change, so I kept this PR lean to make it easier to review the fundamentals of the context manager. Once the context manager becomes mandatory, it can also be used to replace `acquire_spill_lock`, paving the way for much more complete and safer unification of spilling and CoW.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Matthew Roeschke (https://github.com/mroeschke)

URL: https://github.com/rapidsai/cudf/pull/20952